refactor: purchase history separation
